# Game Of Life (React Only)
Max cycle speed of the 50 x 50 square game achievable on my machine is 171 - 200 ms per generation.  At 70 x 100 squares it runs at max speed of approx 380 - 520 ms per generation.

This is using a flat state object. An implementation using an array as a state property makes the array calculation much faster but the render much slower, and the render is the limiting factor.

BUG: Once the size is increased with one of the resize buttons the state contains all the extra properties, so they must be removed. This is annoying and probably slow but it is not a performance issue because it will be rarely performed.