# Rogue Like Game - Fix bug re very slow movement.
Problem was the World component (unchanged by move()) was doing a full re-render (all 10K elements).  Added a shouldComponentUpdate and now movement is very fast as expected.

BUG - At change of levels the call to setEntityLocations is occuring after the attempt to place the entities. Should be a simple fix.