# SVG - Working SVG Mask creating darkness with light circle around player.
GOTCHA.  Had a lot of trouble working out the positioning of the elements and it turned out the cause of my pain
was a div that did not have a position attribute wrapping content that was therefore showing up in front of content
that I expected it to be behind. Check out the previous commits.