#RLG - Refactor animation to use react-addons-css-transition-group and move entity creation settings to settings object
Note the react css transition group is a component that (I think) needs an array of items with their key set to the index in the array. When I changed the key attribute for one of the items, it animated the element with the highest key value, which was not what I wanted.