#RLG - Interesting Bug. Mounting incomplete.
Load the game for the first time it works as expected. Navigate to one of the other apps (in the Contents) and then back to the game and it looks like it has just loaded up as normal, but then the player cannot move.

From the error messages it appears to be a problem with the mounting or unmounting process being incomplete. Possibly due to an unresolved promise.