# RLG Bug 161026 - Router component fails to unmount.
This is probably good because it would suck if you couldn't build state in that method.

Inspecting the React tab of the Developer tools shows that we clearly have two different 'states'.  The `this.state.player` is still the old one (that component must never have dismounted).

At some point `this` must switch from being the component nested in the parent Router component.